Lich skeleton balance slop (#4139)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
As I am on a balance PR streak, I offer the lich skeleton balance slop
PR.
After the removal of infinite stamina (the only thing lich skeleton had
going on for theirselves), lich skellies were left completely useless,
not even able to hold a sword or inflict wounds most of the time.
<img width="1353" height="324" alt="image"
src="https://github.com/user-attachments/assets/bdadb410-d6ab-4130-b4a7-9093e2c3b45d"
/>

-Raise lich skeleton statline.
-Make lich skeleton unable to dodge.

Discord thread:
https://discord.com/channels/748354466335686736/1423817389291733075

## Why It's Good For The Game
The lich is supposed to raise an army of the dead and be a town ending
threat, but what ends up happening is that the lich is the only one
capable of actually fighting in his army, so they just end up being a
one man army and eventually die without being much of a threat.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: Lich skeleton stats raised, lich skeleton can't dodge anymore.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->