Deadite balance slop (#4145)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR tweaks Deadites quite a bit, it is a buff.
This makes deadites slow moving and durable.

Stat changes:
7 STR -> 12 STR
2 SPD -> 1 SPD
5 CON -> 15 CON

Trait changes:
Replaced infinite stamina with infinite energy
Removed critical weakness  (they won't die in 1 crit anymore)
Removed zombie speech (they speak in zizo chant anyways)
removed nolimbdisable

Discord thread: 
https://discord.com/channels/748354466335686736/1423817389291733075

## Why It's Good For The Game
Deadites are supposed to be a very big deal within the setting but an
orphan with a stone knife can kill a deadite if they want to.

People shouldn't let bodies lay around, right now its easier to kill the
deadite that is going to rise than to bury or decapitate them.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: Deadite buffs/tweaks
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->